+++ src/Resources/GPUTexture.cpp
@@ -38,7 +38,7 @@ GPUTexture::GPUTexture(const fs::path& path) {
 		gl->glGenerateTextureMipmap(id);
 		delete data;
 	} else {
-		id = SOIL_load_OGL_texture_from_memory(reader.buffer.data(), reader.buffer.size(), SOIL_LOAD_AUTO, SOIL_LOAD_AUTO, SOIL_FLAG_DDS_LOAD_DIRECT | SOIL_FLAG_SRGB_COLOR_SPACE);
+		id = SOIL_load_OGL_texture_from_memory(reader.buffer.data(), static_cast<int>(reader.buffer.size()), SOIL_LOAD_AUTO, SOIL_LOAD_AUTO, SOIL_FLAG_DDS_LOAD_DIRECT | SOIL_FLAG_SRGB_COLOR_SPACE);
 		if (id == 0) {
 			std::cout << "Error loading texture: " << path << "\n";
 		}
