+++ src/Resources/SkinnedMesh/SkeletalModelInstance.cpp
@@ -39,7 +39,7 @@ SkeletalModelInstance::SkeletalModelInstance(std::shared_ptr<mdx::MDX> model) : 
 
 	for (size_t i = 0; i < model->sequences.size(); i++) {
 		if (model->sequences[i].name.find("Stand") != std::string::npos) {
-			set_sequence(i);
+			set_sequence(static_cast<int>(i));
 			break;
 		}
 	}
